refactor: Category Exception 패키지 구조 변경
